Report an error on a negative length for bytes or string
This change fixes the crash in the last commit. After upgrading to 1.4
nightly, the code panicked instead of crashing, revealing that Vec was
being passed a huge capacity (due to a negative length being cast).
Vec evidently had a bug that it did not validate capacity.

Now the incorrect test with the negative length is fixed, and a test has
been added for checking for a negative length.